move troubleshooting_info to its own module
