+++ nfl_data_py/__init__.py
@@ -1154,18 +1154,3 @@ def clean_nfl_data(df):
         df.replace({'col_team': col_tm_repl}, inplace=True)
 
     return df
-
-
-def troubleshooting_info():
-    """Print system, package and deps information relevant to troubleshooting"""
-    import platform as pf
-    import importlib.metadata as md
-
-    print("nfl_data_py", md.version("nfl_data_py"))
-    print(pf.python_implementation(), pf.python_version())
-    print(pf.platform())
-    print("Dependencies:")
-    for pkg in md.requires("nfl_data_py"):
-        name, constraints = pkg.split(maxsplit=1)
-        version = f"v{md.version(name).strip('.0')}"
-        print(f" - {name}\t{version}\t({constraints})")
